+++ packages/svelte/tests/compiler-errors/samples/runes-wrong-state-snapshot-args/_config.js
@@ -1,0 +1,8 @@
+import { test } from '../../test';
+
+export default test({
+	error: {
+		code: 'invalid-rune-args-length',
+		message: '$state.snapshot can only be called with 1 argument'
+	}
+});