+++ packages/svelte/tests/compiler-errors/samples/runes-wrong-state-snapshot-args/main.svelte
@@ -1,0 +1,3 @@
+<script>
+	const foo = $state.snapshot();
+</script>