+++ packages/svelte/tests/runtime-runes/samples/unstate/main.svelte
@@ -1,7 +1,0 @@
-<script>
-		import { unstate } from 'svelte';
-
-    let items = $state([{a: 0}]);
-</script>
-
-<button on:click={() => items.push({a: items.length})}>{JSON.stringify(structuredClone(unstate(items)))}</button>