+++ packages/svelte/CHANGELOG.md
@@ -1,5 +1,13 @@
 # svelte
 
+## 5.0.0-next.105
+
+### Patch Changes
+
+- breaking: remove unstate(), replace with $state.snapshot rune ([#11180](https://github.com/sveltejs/svelte/pull/11180))
+
+- fix: more accurate default value handling ([#11183](https://github.com/sveltejs/svelte/pull/11183))
+
 ## 5.0.0-next.104
 
 ### Patch Changes